requirements.txt
imagehash
salesforce-lavis
pandas
huggingface-hub
Pillow
onnxruntime-gpu
wordcloud
openai
gradio
--index-url https://download.pytorch.org/whl/cu118
torch==2.2.1+cu118
torchvision==0.17.1+cu118
torchaudio==2.2.1+cu118
numpy==1.26.4
ultralytics==8.1.47
opencv-python==4.8.0.74